✨ Add yarn script to build iOS (#18)
Xcode has decided to start enforcing caps in the name of schemes. So
when `react-native run-ios` was called, the scheme apprenticeprogress
was looked for rather than ApprenticeProgress.

Rather than dealing with this annoyance closer to the root (which could
have side effects), I’m adding this yarn command which specifies the
scheme name with capitals - and to call an iPhone X while I'm at it.