✨ Add login (#22)
A barebones login feature. A user can login and be provided with a token that's sent with proceeding API requests. On logout, the token is cleared from the local cache.

The current flow-typed definition for react navigation isn’t written very well for my needs so the getParam and navigate overwrites are needed to not get any errors.

Major:
* ✨ Add Login screen
* ✨ Add login to navigation
* ✨ Add FormLogin component
* ✨ Add authentication flow between screens
* ♻️ Refactor client for authentication token

Minor:
* ♻️ Refactor handlePress into utility function
* ✨ Add loginMutation query to get token from API
* ♻️ Refactor Login to auto-login if token present
* ♻️ Refactor Navigation type
* 🔧 Hide flow-type folder from PR display